feat: filter SQL and skip empty deltas.
logion-network/logion-internal#423

+++ src/BackupManager.ts
@@ -54,49 +54,58 @@ export class BackupManager {
         let backupFile: BackupFileName;
         let backupFilePath: string;
         const lastFullBackup = journal.getLastFullBackup();
-        let logsToRemove: string[] = [];
+        let deltaBackupResult: DeltaBackupResult | undefined = undefined;
         if(lastFullBackup === undefined
                 || date.diff(lastFullBackup.fileName.date) > this.configuration.maxDurationSinceLastFullBackup) {
+            logger.info("Producing full backup...");
             backupFile = BackupFileName.getFullBackupFileName(date);
             backupFilePath = path.join(this.configuration.workingDirectory, backupFile.fileName);
             await this.doFullBackup(backupFilePath);
         } else {
+            logger.info("Producing delta...");
             backupFile = BackupFileName.getDeltaBackupFileName(date);
             backupFilePath = path.join(this.configuration.workingDirectory, backupFile.fileName);
-            logsToRemove = await this.doDeltaBackup(backupFilePath);
+            deltaBackupResult = await this.doDeltaBackup(backupFilePath);
         }
 
-        logger.info(`Adding file ${backupFilePath} to IPFS...`);
-        const cid = await this.configuration.fileManager.moveToIpfs(backupFilePath);
-        logger.info(`Success, file ${backupFilePath} got CID ${cid}`);
+        if(deltaBackupResult === undefined || !deltaBackupResult.emptyDelta) {
+            logger.info(`Adding file ${backupFilePath} to IPFS...`);
+            const cid = await this.configuration.fileManager.moveToIpfs(backupFilePath);
+            logger.info(`Success, file ${backupFilePath} got CID ${cid}`);
 
-        journal.addBackup(new BackupFile({cid, fileName: backupFile}));
+            journal.addBackup(new BackupFile({cid, fileName: backupFile}));
 
-        const toRemove = journal.keepOnlyLastFullBackups(this.configuration.maxFullBackups);
-        for(const file of toRemove) {
-            this.configuration.fileManager.removeFileFromIpfs(file.cid);
+            const toRemove = journal.keepOnlyLastFullBackups(this.configuration.maxFullBackups);
+            for(const file of toRemove) {
+                this.configuration.fileManager.removeFileFromIpfs(file.cid);
+            }
+
+            logger.info("Writing journal...");
+            await journal.write(this.configuration.journalFile);
+            logger.info("Journal successfully written, sending by e-mail...");
+
+            await this.configuration.mailer.sendMail({
+                to: this.configuration.mailTo,
+                subject: "Backup journal updated",
+                text: "New journal file available, see attachment.",
+                attachments: [
+                    {
+                        path: this.configuration.journalFile,
+                        filename: "journal.txt"
+                    }
+                ]
+            });
+        } else {
+            logger.info("No change detected.");
         }
 
-        logger.info("Writing journal...");
-        await journal.write(this.configuration.journalFile);
-        logger.info("Journal successfully written, sending by e-mail...");
-
-        await this.configuration.mailer.sendMail({
-            to: this.configuration.mailTo,
-            subject: "Backup journal updated",
-            text: "New journal file available, see attachment.",
-            attachments: [
-                {
-                    path: this.configuration.journalFile,
-                    filename: "journal.txt"
-                }
-            ]
-        });
-
-        logger.info("Removing processed logs...");
-        for(const file of logsToRemove) {
-            this.configuration.fileManager.deleteFile(file);
+        if(deltaBackupResult !== undefined) {
+            logger.info("Removing processed logs...");
+            for(const file of deltaBackupResult.logsToRemove) {
+                this.configuration.fileManager.deleteFile(file);
+            }
         }
+
         logger.info("All done.");
     }
 
@@ -121,32 +130,53 @@ export class BackupManager {
         }
     }
 
-    private async doDeltaBackup(backupFilePath: string): Promise<string[]> {
+    private async doDeltaBackup(backupFilePath: string): Promise<DeltaBackupResult> {
         const writer = new EncryptedFileWriter(this.configuration.password);
         await writer.open(backupFilePath);
         try {
-            const toRemove: string[] = [];
+            const logsToRemove: string[] = [];
+            let emptyDelta = true;
             const logsProcessor = new LogsProcessor({
                 sqlSink: async (sql) => {
-                    if(sql) {
-                        await writer.write(Buffer.from(sql, 'utf-8'));
+                    const filteredSql = this.filterSql(sql);
+                    if(filteredSql) {
+                        emptyDelta = false;
+                        await writer.write(Buffer.from(filteredSql, 'utf-8'));
                     } else {
                         return Promise.resolve();
                     }
                 },
                 filePostProcessor: (file: string) => {
-                    toRemove.push(file);
+                    logsToRemove.push(file);
                     return Promise.resolve();
                 }
             });
             await logsProcessor.process(this.configuration.logDirectory);
             await writer.close();
-            return toRemove;
+
+            if(emptyDelta) {
+                await this.configuration.fileManager.deleteFile(backupFilePath);
+            }
+
+            return { logsToRemove, emptyDelta };
         } catch(e) {
             await writer.close();
             await this.configuration.fileManager.deleteFile(backupFilePath);
             throw e;
         }
+    }
+
+    private filterSql(sql: string | undefined): string | undefined {
+        if(sql
+            && this.isNotSyncPointUpdate(sql)) {
+            return sql;
+        } else {
+            return undefined;
+        }
+    }
+
+    private isNotSyncPointUpdate(sql: string): boolean {
+        return !sql.startsWith('UPDATE "sync_point" SET "latest_head_block_number" =');
     }
 }
 
@@ -162,4 +192,9 @@ class PgDumpProcessHandler extends ProcessHandler {
     async onStdOut(data: any) {
         await this.writer.write(data);
     }
+}
+
+interface DeltaBackupResult {
+    logsToRemove: string[];
+    emptyDelta: boolean;
 }
